requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.33.0
coreschema==0.0.4
coreapi==2.3.3
django==4.2.1
django-recaptcha==3.0.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.2
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==23.1
psycopg2-binary==2.9.6
pyjwt==2.7.0
pytz==2023.3
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
setuptools==67.7.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==1.26.15
